good work but not finished

+++ ft_strchr.c
@@ -6,22 +6,22 @@
 /*   By: yallo <yallo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 14:33:31 by yallo             #+#    #+#             */
-/*   Updated: 2022/11/15 15:32:40 by yallo            ###   ########.fr       */
+/*   Updated: 2022/11/18 14:18:50 by yallo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char *ft_strchr(char *s ,int c)
+char	*ft_strchr(char *s, int c)
 {
-    int i;
+	size_t	i;
 
-    i = 0;
-    while (i < strlen(s) + 1)
-    {
-        if (s[i] == c)
-            return(s+i);
-        i++;
-    }
-    return (NULL);
+	i = 0;
+	while (i < strlen(s) + 1)
+	{
+		if (s[i] == c)
+			return (s + i);
+		i++;
+	}
+	return (NULL);
 }
